+++ services/src/main/java/org/keycloak/services/clientpolicy/condition/ClientAccessTypeConditionFactory.java
@@ -21,15 +21,13 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import org.keycloak.Config.Scope;
 import org.keycloak.models.KeycloakSession;
-import org.keycloak.models.KeycloakSessionFactory;
 import org.keycloak.provider.ProviderConfigProperty;
 
 /**
  * @author <a href="mailto:takashi.norimatsu.ws@hitachi.com">Takashi Norimatsu</a>
  */
-public class ClientAccessTypeConditionFactory implements ClientPolicyConditionProviderFactory {
+public class ClientAccessTypeConditionFactory extends AbstractClientPolicyConditionProviderFactory {
 
     public static final String PROVIDER_ID = "client-access-type";
 
@@ -42,6 +40,8 @@ public class ClientAccessTypeConditionFactory implements ClientPolicyConditionPr
     private static final List<ProviderConfigProperty> configProperties = new ArrayList<ProviderConfigProperty>();
 
     static {
+        addCommonConfigProperties(configProperties);
+
         ProviderConfigProperty property;
         property = new ProviderConfigProperty(TYPE, "client-accesstype.label", "client-accesstype.tooltip", ProviderConfigProperty.MULTIVALUED_LIST_TYPE, TYPE_CONFIDENTIAL);
         List<String> updateProfileValues = Arrays.asList(TYPE_CONFIDENTIAL, TYPE_PUBLIC, TYPE_BEARERONLY);
@@ -52,18 +52,6 @@ public class ClientAccessTypeConditionFactory implements ClientPolicyConditionPr
     @Override
     public ClientPolicyConditionProvider create(KeycloakSession session) {
         return new ClientAccessTypeCondition(session);
-    }
-
-    @Override
-    public void init(Scope config) {
-    }
-
-    @Override
-    public void postInit(KeycloakSessionFactory factory) {
-    }
-
-    @Override
-    public void close() {
     }
 
     @Override
@@ -80,5 +68,4 @@ public class ClientAccessTypeConditionFactory implements ClientPolicyConditionPr
     public List<ProviderConfigProperty> getConfigProperties() {
         return configProperties;
     }
-
 }
