+++ services/src/main/java/org/keycloak/services/clientpolicy/condition/ClientUpdaterContextConditionFactory.java
@@ -21,15 +21,13 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import org.keycloak.Config.Scope;
 import org.keycloak.models.KeycloakSession;
-import org.keycloak.models.KeycloakSessionFactory;
 import org.keycloak.provider.ProviderConfigProperty;
 
 /**
  * @author <a href="mailto:takashi.norimatsu.ws@hitachi.com">Takashi Norimatsu</a>
  */
-public class ClientUpdaterContextConditionFactory implements ClientPolicyConditionProviderFactory {
+public class ClientUpdaterContextConditionFactory extends AbstractClientPolicyConditionProviderFactory {
 
     public static final String PROVIDER_ID = "client-updater-context";
 
@@ -43,6 +41,8 @@ public class ClientUpdaterContextConditionFactory implements ClientPolicyConditi
     private static final List<ProviderConfigProperty> configProperties = new ArrayList<ProviderConfigProperty>();
 
     static {
+        addCommonConfigProperties(configProperties);
+
         ProviderConfigProperty property;
         property = new ProviderConfigProperty(UPDATE_CLIENT_SOURCE, "Update Client Context", "Specifies the context how is client created or updated. " +
                 "ByInitialAccessToken is usually OpenID Connect client registration with the initial access token. " +
@@ -60,18 +60,6 @@ public class ClientUpdaterContextConditionFactory implements ClientPolicyConditi
     }
 
     @Override
-    public void init(Scope config) {
-    }
-
-    @Override
-    public void postInit(KeycloakSessionFactory factory) {
-    }
-
-    @Override
-    public void close() {
-    }
-
-    @Override
     public String getId() {
         return PROVIDER_ID;
     }
@@ -85,5 +73,4 @@ public class ClientUpdaterContextConditionFactory implements ClientPolicyConditi
     public List<ProviderConfigProperty> getConfigProperties() {
         return configProperties;
     }
-
 }
